refactor: 충돌처리
